Part

+++ src/service.rs
@@ -1,11 +1,11 @@
 use cgmath::{Vector3, Array};
 
-pub fn pos_to_index(pos: Vector3<f32>, side_len: i32, ) -> u32 {
+pub fn pos_to_index(pos: Vector3<f32>, side_len: i32, ) -> usize {
     Vector3 {
         x: ((pos.x as i32) % side_len),
             y: ((pos.y as i32) % side_len) * side_len * side_len,
                 z: ((pos.z as i32) % side_len) * side_len
-    }.sum() as u32
+    }.sum() as usize
 }
 
 pub fn index_to_pos(index: u32, side_len: u32, ) -> Vector3<f32> {
